requirements.txt
importlib_resources
allennlp>=0.8.2;python_version  > '3.0'
editdistance
future;python_version < '3.0'
lark-parser
nltk
pandas
xmltodict
yieldfrom